show menu in compact mode
